Support to continue generating data from stopped point in case of environment stopping/ failing

+++ dataset_generator.py
@@ -8,7 +8,7 @@ import numpy as np
 import os
 import contextlib
 
-PARENT_FOLDER = '/Users/danielleyaffe/GitHub/MultiChannelGeneration'
+PARENT_FOLDER = '/gpfs0/bgu-br/users/yaffedan/MultiChannelAudioGenerater'
 AV_DATA_FOLDER = os.path.join(PARENT_FOLDER, 'cleandata')
 A_DATA_FOLDER = os.path.join(PARENT_FOLDER, 'cleanaudiodata')
 
@@ -45,7 +45,7 @@ if __name__ == "__main__":
         num_interfering_sources = 1  # Number of interferes, choose >4 for babble noise
         noise_name = str(num_interfering_sources)+'Inter'
     elif noise == 'babble':
-        num_interfering_sources = 4  # Number of interferes, choose >3 for babble noise
+        num_interfering_sources = 3  # Number of interferes, choose >3 for babble noise
         noise_name = 'Babble'
     else:
         raise ValueError('Invalid noise request. Please choose a noise type: interfere or babble')
@@ -54,13 +54,29 @@ if __name__ == "__main__":
     clean_output_dir = output_dir+'_clean'
     noise_output_dir = output_dir+'_noise'
 
-    if os.path.isfile(f"{output_dir}.json"):
-        # Prevent appending to existing file new dataset generation!
-        os.remove(f"{output_dir}.json")
-        print(f"Removed existing metadata file: {output_dir}.json")
 
-    # Create list of audio pairs/ combination for simulations.
-    speaker_pairs = generate_speaker_pairs(datapath=A_DATA_FOLDER, num_samples=num_pairs_per_spk, num_interferers=num_interfering_sources, max_total_samples=max_total_samples)
+
+    # Create list of audio pairs/ combination for simulations. If a list exist reload it.
+    if not os.path.exists(os.path.join(PARENT_FOLDER, "speaker_pairs.json")):
+        speaker_pairs = generate_speaker_pairs(datapath=A_DATA_FOLDER, num_samples=num_pairs_per_spk, num_interferers=num_interfering_sources, max_total_samples=max_total_samples)
+        if os.path.isfile(f"{output_dir}.json"):
+            # Prevent appending to existing file new dataset generation!
+            os.remove(f"{output_dir}.json")
+            print(f"Removed existing metadata file: {output_dir}.json")
+    else:
+        with open(os.path.join(PARENT_FOLDER, "speaker_pairs.json"), "r") as f:
+            speaker_pairs = json.load(f)
+        print(f"Loaded existing speaker pairs from {os.path.join(PARENT_FOLDER, 'speaker_pairs.json')}. {len(speaker_pairs)} existing pairs")
+
+        sample_ids = set()  # use a set for quick membership checks
+
+        with open(f"{output_dir}.json", "r") as f:
+            for line in f:
+                if line.strip():  # skip empty lines
+                    entry = json.loads(line)
+                    sample_ids.add(entry["sample_id"])
+        print(f"Existing metadata file: {output_dir}.json will be used/updated.")
+
     dataset_generator = MultiChannelGenerator(sample_rate=16000,
                                               output_dir=output_dir,
                                               clean_output_dir=clean_output_dir,
@@ -74,43 +90,54 @@ if __name__ == "__main__":
     idx = 0
     metadata = []
     with open('log_data_generator.log', 'w') as log_file:
-        with contextlib.redirect_stdout(log_file):
-            with tqdm(total=len(speaker_pairs), desc="Generating dataset", unit="sample") as pbar:
-                while idx < len(speaker_pairs):
-                    snr, room_dim, rt60_tgt = random.choice(snr_values), random.choice(room_sizes), random.choice(rt60_values)
+        log_file.write(f"Starting dataset generation with simulation: {simulation}, noise: {noise}, SNR values: {snr_values}, room sizes: {room_sizes}, RT60 values: {rt60_values}\n")
+        log_file.flush()
+        with tqdm(total=len(speaker_pairs), desc="Generating dataset", unit="sample") as pbar:
+            while idx < len(speaker_pairs):
+                snr, room_dim, rt60_tgt = random.choice(snr_values), random.choice(room_sizes), random.choice(rt60_values)
 
-                    audio_files = speaker_pairs[indexes[idx]]
-                    if idx==len(speaker_pairs)-1:
-                        dataset_generator.verbose = True
-                        dataset_generator.verbose_outpur_dir = output_dir+'_plots'
-                        os.makedirs(output_dir+'_plots', exist_ok=True)
-                    try:
-                        sample_meta_data = dataset_generator.generate_multichannel_audio(room_dim=room_dim,
-                                                                                            rt60_tgt=rt60_tgt,
-                                                                                            snr=snr,
-                                                                                            audio_files=audio_files,
-                                                                                            num_interfering_sources=num_interfering_sources,
-                                                                                            with_DRR=True,
-                                                                                            save_audio=save_multichannel,
-                                                                                            save_noise_audio=save_noise_audio)
-                        if save_binaural:
-                            dataset_generator.generate_binaural_audio(room_dim, rt60_tgt, snr, audio_files, num_interfering_sources,
-                                                                azimuth_deg=90.0, save_audio=save_binaural)
-                        if save_multichannel:
-                            with open(f"{output_dir}.json", "a") as f:
-                                json.dump(sample_meta_data, f)  # Dump each dictionary separately
-                                f.write("\n")  # Add a newline after each JSON object
-                        if verbose:
-                            log_file.write(sample_meta_data + '\n')
-
-                        metadata.append(sample_meta_data)
-                        idx += 1
-                    
-                    except Exception as e:
-                        log_file.write("\n{ Error while processing: " + str(audio_files) + '\n' + str(e) + '}\n')
-                        print('Error while processing: ' + str(audio_files) + 'see log_reverb.log')
-    
+                audio_files = speaker_pairs[indexes[idx]]
+                
+                if audio_files["ID"] in sample_ids:
+                    # print(f"Sample ID {audio_files['ID']} already exists in metadata. Skipping generation.")
+                    log_file.write(f"Sample ID {audio_files['ID']} already exists in metadata. Skipping generation.")
+                    log_file.flush()
+                    idx += 1
                     pbar.update(1)
+                    continue
+
+                if idx==len(speaker_pairs)-1:
+                    dataset_generator.verbose = True
+                    dataset_generator.verbose_outpur_dir = output_dir+'_plots'
+                    os.makedirs(output_dir+'_plots', exist_ok=True)
+                try:
+                    sample_meta_data = dataset_generator.generate_multichannel_audio(room_dim=room_dim,
+                                                                                        rt60_tgt=rt60_tgt,
+                                                                                        snr=snr,
+                                                                                        audio_files=audio_files,
+                                                                                        num_interfering_sources=num_interfering_sources,
+                                                                                        with_DRR=True,
+                                                                                        save_audio=save_multichannel,
+                                                                                        save_noise_audio=save_noise_audio)
+                    if save_binaural:
+                        dataset_generator.generate_binaural_audio(room_dim, rt60_tgt, snr, audio_files, num_interfering_sources,
+                                                            azimuth_deg=90.0, save_audio=save_binaural)
+                    if save_multichannel:
+                        with open(f"{output_dir}.json", "a") as f:
+                            json.dump(sample_meta_data, f)  # Dump each dictionary separately
+                            f.write("\n")  # Add a newline after each JSON object
+                    if verbose:
+                        log_file.write(sample_meta_data + '\n')
+
+                    metadata.append(sample_meta_data)
+                    
+                
+                except Exception as e:
+                    log_file.write("\n{ Error while processing: " + str(audio_files) + '\n' + str(e) + '}\n')
+                    print('Error while processing: ' + str(audio_files) + 'see log_reverb.log')
+
+                idx += 1
+                pbar.update(1)
                 
 
     print(f"Dataset generation complete! {idx} samples successfully processed. Save status: Multi channel:{save_multichannel}, Multi Channel Noise:{save_noise_audio}, Binaural:{save_binaural}")
